support underscore in numbers
A while back Go added underscores to the syntax of numbers (for example
3_000). The package go/scanner automatically supports this but
strconv.Atoi does not, resulting in an error when they are used.
The goval documentation doesn't specify the syntax of its numbers but I
think it would be useful if it did support underscores, for the same
reasons they were added to Go.